refactor: Extract the app core from its skeleton
We would like to have a stand-alone application
that can be used as a library.
So we have here dissociate the code that allows to build an application,
from the code of the application itself.

There is no modification in the code.
