+++ src/components/App.spec.jsx
@@ -1,55 +1,32 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
-import { usePapersDefinitions } from 'src/components/Hooks/usePapersDefinitions'
 import { App } from 'src/components/App'
+import AppLike from 'test/components/AppLike'
 
 /* eslint-disable react/display-name */
-jest.mock('src/components/Contexts/FormDataProvider', () => ({
-  FormDataProvider: ({ children }) => (
-    <div data-testid="FormDataProvider">{children}</div>
-  )
+jest.mock('src/indexLib', () => () => <div data-testid="indexLib" />)
+jest.mock('cozy-ui/transpiled/react/Layout', () => ({
+  Layout: ({ children }) => <div data-testid="Layout">{children}</div>,
+  Main: ({ children }) => <div data-testid="Main">{children}</div>,
+  Content: ({ children }) => <div data-testid="Content">{children}</div>
 }))
-jest.mock('src/components/StepperDialog/StepperDialogWrapper', () => () => (
-  <div data-testid="StepperDialogWrapper" />
+jest.mock('cozy-ui/transpiled/react/Icon/Sprite', () => () => (
+  <div data-testid="Sprite" />
 ))
-jest.mock('src/components/AppLayout', () => ({
-  AppLayout: ({ children }) => <div data-testid="AppLayout">{children}</div>
-}))
-jest.mock('src/components/AppRouter', () => ({
-  AppRouter: () => <div data-testid="AppRouter" />
-}))
-jest.mock('src/components/Hooks/useStepperDialog')
-jest.mock('src/components/Hooks/usePapersDefinitions')
 /* eslint-enable react/display-name */
 
 describe('App', () => {
-  it('should contain progressbar when no papers', () => {
-    useStepperDialog.mockReturnValue({ isStepperDialogOpen: false })
-    usePapersDefinitions.mockReturnValue({ papersDefinitions: [] })
+  it('should contain Layout, Main, Content & Sprite components', () => {
+    const { queryByTestId } = render(
+      <AppLike>
+        <App />
+      </AppLike>
+    )
 
-    const { getByRole } = render(<App />)
-
-    expect(getByRole('progressbar')).toBeTruthy()
-  })
-
-  it('should return AppRouter component', () => {
-    useStepperDialog.mockReturnValue({ isStepperDialogOpen: false })
-    usePapersDefinitions.mockReturnValue({ papersDefinitions: ['1', '2'] })
-
-    const { queryByTestId } = render(<App />)
-
-    expect(queryByTestId('AppRouter')).toBeTruthy()
-  })
-
-  it('should return FormDataProvider & StepperDialogWrapper component', () => {
-    useStepperDialog.mockReturnValue({ isStepperDialogOpen: true })
-    usePapersDefinitions.mockReturnValue({ papersDefinitions: [] })
-
-    const { queryByTestId } = render(<App />)
-
-    expect(queryByTestId('FormDataProvider')).toBeTruthy()
-    expect(queryByTestId('StepperDialogWrapper')).toBeTruthy()
+    expect(queryByTestId('Layout')).toBeTruthy()
+    expect(queryByTestId('Main')).toBeTruthy()
+    expect(queryByTestId('Content')).toBeTruthy()
+    expect(queryByTestId('Sprite')).toBeTruthy()
   })
 })
